Started V2
This tool was written in 2014 when I did not know that much this update is here to pretty much rewrite the whole thing to make it better. This commit is a complete rework of the UI used for script forge along with using a Scriptable Singleton pattern for saving data.