New Error System and Hash Generation
Errors message are now based off an enum an allow toggling off select types. 
All forges now uses a hash behind the scenes which will stop them from building if the hash is the same is last time. This takes into account of the file is missing and forces a rebuild anyways.